+++ examples/physics/demo_physics_tour.py
@@ -4,7 +4,7 @@ Tour of forward sensing operators
 
 This example provides a tour of some of the forward operators implemented in DeepInverse.
 We restrict ourselves to operators where the signal is a 2D image. The full list of operators can be found in
-`here <physics>`_.
+:ref:`here <physics>`.
 
 """
 
@@ -12,7 +12,7 @@ import torch
 
 import deepinv as dinv
 from deepinv.utils.plotting import plot
-from deepinv.utils.demo import load_example
+from deepinv.utils import load_example
 
 
 # %%