+++ CHANGELOG.rst
@@ -12,10 +12,14 @@ New Features
 - Add dataset for patch sampling from (large) nD images without loading entire images into memory (:gh: `806` by `Vicky De Ridder`_)
 - Add support for 3D CNNs (:gh: `869` by `Vicky De Ridder`)
 - Add support for complex dtypes in WaveletDenoiser, WaveletDictDenoiser and WaveletPrior (:gh:`738` by `Chaithya G R`_)
+- dinv.io functions for loading DICOM, NIFTI, COS, GEOTIFF etc. (:gh:`768` by `Andrew Wang`_)
 - Add support for Poisson2Sparse (:gh:`677` by `Jérémy Scanvic`_)
 
 Changed
 ^^^^^^^
+- load_np_url now returns tensors, load_url helper function moved to io (:gh:`768` by `Andrew Wang`_)
+- utils/signal.py renamed to signals.py to avoid stdlib conflict (:gh:`768` by `Andrew Wang`_)
+- utils.get_data_home now creates folder if not exist (:gh:`768` by `Andrew Wang`_)
 
 Fixed
 ^^^^^